pyproject.toml
[tool.poetry]
name = "llm-agents"
version = "0.0.1"
description = ""
authors = ["Sharabh Shukla <sharabh2019@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
langchain-community = "^0.0.17"
duckduckgo-search = "^4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
